Move query implementation to sep module
Adds:
- unit tests
- fixes default value of negate
