+++ src/obr/create_tree.py
@@ -38,10 +38,12 @@ def flatten(d, parent_key="", sep="/"):
 
 def get_path_from(operation: dict, value) -> str:
     if not operation.get("schema"):
-        print("""Error Schema missing for
+        print(
+            """Error Schema missing for
 
 Set schema to allow creating views
-        """)
+        """
+        )
         raise KeyError
 
     return operation["schema"].format(**flatten(value)) + "/"
